Separating footer from portfolio and updating site layout
