+++ shared/meta/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./lib/async-wrapper/async-wrapper";
+export { default as Footer } from "./lib/footer/footer";
 export { default as Layout } from "./lib/layout/layout";
 export { default as AsyncWrapper } from "./lib/async-wrapper/async-wrapper";
 export { default as Sitenav } from "./lib/sitenav/sitenav";
