+++ shared/meta/src/lib/footer/footer.data.ts
@@ -1,0 +1,9 @@
+import { ContactFields } from "@erkjbro-tech/shared/apis";
+
+export const LINK_DATA: ContactFields[] = [
+  { title: "", href: "mailto:erkjbro@erikjbrown.tech", text: "Email" },
+  { title: "", href: "https://www.linkedin.com/in/erkjbro/", text: "LinkedIn" },
+  { title: "", href: "https://github.com/erkjbro", text: "GitHub" },
+  { title: "", href: "https://twitter.com/erkjbro", text: "Twitter" },
+  { title: "", href: "https://www.upwork.com/fl/erkjbro", text: "Upwork" }
+];
