+++ shared/meta/src/lib/footer/footer.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import Footer from './footer';
+
+describe('Footer', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<Footer />);
+    expect(baseElement).toBeTruthy();
+  });
+});
