+++ shared/meta/src/lib/layout/layout.styled.tsx
@@ -1,6 +1,13 @@
 import styled from "styled-components";
 
 export const StyledLayout = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
   text-align: center;
-  height: 100vh;
+  min-height: 100vh;
+`;
+
+export const LayoutBody = styled.div`
+  min-height: calc(100vh - 14rem);
 `;
