+++ src/app/portfolio/portfolio-store.ts
@@ -3,13 +3,13 @@ import { useEffect, useReducer, useState } from "react";
 
 import { FetchStatus } from "@erkjbro-tech/shared/meta";
 import { ContentfulApi, ContentTypes } from "@erkjbro-tech/shared/apis";
-import type { ContactFields, StaticPageFields } from "@erkjbro-tech/shared/apis";
+import type { StaticPageFields } from "@erkjbro-tech/shared/apis";
 
 const SPACE_ID: string = import.meta.env.VITE_CONTENT_SPACE_ID;
 const API_KEY: string = import.meta.env.VITE_CONTENT_API_KEY;
 
 // TODO: Convert to a static page store for an api lib component.
-export type PortfolioStoreTuple = [FetchStatus, StaticPageFields, ContactFields[]];
+export type PortfolioStoreTuple = [FetchStatus, StaticPageFields];
 
 const usePortfolioStore = (): PortfolioStoreTuple => {
   const INITIAL_STATE: StaticPageFields = {
@@ -17,15 +17,6 @@ const usePortfolioStore = (): PortfolioStoreTuple => {
     body: "",
     header: "Erik J Brown"
   };
-
-  // TODO: Move to a footer component (or someplace else), and fetch from Contentful.
-  const LINK_DATA: ContactFields[] = [
-    { title: "", href: "mailto:erkjbro@erikjbrown.tech", text: "Email" },
-    { title: "", href: "https://www.linkedin.com/in/erkjbro/", text: "LinkedIn" },
-    { title: "", href: "https://github.com/erkjbro", text: "GitHub" },
-    { title: "", href: "https://twitter.com/erkjbro", text: "Twitter" },
-    { title: "", href: "https://www.upwork.com/fl/erkjbro", text: "Upwork" }
-  ];
 
   type ActionTypes = "SET_PORTFOLIO";
   type Actions = {
@@ -78,7 +69,7 @@ const usePortfolioStore = (): PortfolioStoreTuple => {
     })();
   }, []);
 
-  return [status, state, LINK_DATA];
+  return [status, state];
 };
 
 export default usePortfolioStore;
