+++ src/app/portfolio/portfolio.styled.tsx
@@ -32,49 +32,3 @@ export const PortfolioContent = styled.div`
     }
   }
 `;
-
-export const ContactList = styled.ul`
-  height: fit-content;  
-  margin: 1.2rem 0.5rem;
-  border-top: 2px solid #11114a;
-  padding: 1rem 0.2rem 0.5rem 0.2rem;
-  display: flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  align-items: center;
-  text-align: left;
-  
-
-  @media (min-width: 480px) {
-    flex-direction: row;
-    text-align: center;
-  }
-  
-  li {
-    display: inherit;
-    align-content: center;
-  }
-  
-  li a {
-    width: 100px;    
-    display: inherit;    
-    padding: 0.2rem 0.4rem;
-    margin: 0.2rem 0;    
-    border: 2px solid transparent;
-    align-items: center;
-    color: #11114a;
-    font-size: 1.2rem;
-    font-weight: bold;
-    text-decoration: none;
-    
-    &:hover, &:focus {
-      border: 2px solid #11114a;
-      border-radius: 10px;
-    }
-    
-    @media (min-width: 480px) {
-      width: fit-content;
-    }
-  }
-`;
